refactor:#191 import 문 자동 수정 안 된 부분 수정 반영
