+++ src/app/(checklist)/checklist/[mission]/page.tsx
@@ -2,7 +2,11 @@ import { notFound } from 'next/navigation';
 import ChecklistClient from '@/components/features/checklist/checklist-client';
 import { validMissionTags } from '@/constants/mission';
 import { getUserSessionState } from '@/lib/utils/api/auth/auth.api';
-import { getCompletedMissionIds, getMissionListByLevel, getUserLevelByMission } from '@/lib/utils/api/checklist.api';
+import {
+  getCompletedMissionIds,
+  getMissionListByLevel,
+  getUserLevelByMission
+} from '@/lib/utils/api/checklist/checklist.api';
 import type { Level, MissionTag } from '@/types/checklist';
 
 let userLevel = '1'; // default level (for 비로그인 사용자)
