+++ src/components/features/mypage/button-client-component.tsx
@@ -3,8 +3,8 @@
 import Image from 'next/image';
 import { useState } from 'react';
 import RandomMissionModal from '@/components/features/modals/random-mission';
-import CHEVRON_RIGHT from '@images/images/chevron-right.svg';
 import type { TableMissionList } from '@/types/supabase-const';
+import CHEVRON_RIGHT from '@images/images/chevron-right.svg';
 
 interface ButtonClientComponentProps {
   missionsData: TableMissionList[];
