+++ src/lib/hooks/queries/use-get-life-post-by-mission-id.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
-import { getLifePostByMissionId } from '@/lib/utils/api/life-by-mission.api';
+import { getLifePostByMissionId } from '@/lib/utils/api/life/life-by-mission.api';
 import type { LifePostWithImageUrls } from '@/types/life-post';
 
 export const useGetLifePostByMissionId = (missionId: number | null) => {
