+++ src/lib/hooks/queries/use-life-posts-by-month-range.ts
@@ -2,7 +2,7 @@
 
 import { useQueries } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
-import { getLifePostsByMonth } from '@/lib/utils/api/life-api-client';
+import { getLifePostsByMonth } from '@/lib/utils/api/life/life-api-client';
 import type { LifePostWithImageUrls } from '@/types/life-post';
 
 export const useLifePostsByMonthRange = (months: string[]) => {
