+++ src/lib/hooks/queries/use-life-posts-by-month.ts
@@ -1,7 +1,7 @@
 // lib/hooks/queries/useLifePostsByMonth.ts
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
-import { getLifePostsByMonth } from '@/lib/utils/api/life-api-client';
+import { getLifePostsByMonth } from '@/lib/utils/api/life/life-api-client';
 import type { LifePostWithImageUrls } from '@/types/life-post';
 
 export const useLifePostsByMonth = (month: string) => {
