+++ src/lib/utils/api/checklist/checklist-post.api.ts
@@ -1,8 +1,8 @@
 import { MSG } from '@/constants/messages';
 import { TABLE } from '@/constants/supabase-tables-name';
-import { getCompletedMissionIds, getMissionListByLevel } from '@/lib/utils/api/checklist.api';
 import { createClient } from '@/lib/utils/supabase/supabase-server';
 import type { MissionTag, MissionType } from '@/types/checklist';
+import { getCompletedMissionIds, getMissionListByLevel } from '@/lib/utils/api/checklist/checklist.api';
 
 interface DropdownParams {
   userId: string;
